state: started moving multi-env utilities to a new file
A number of other functions will be added here.
